Fix the parsing error of Issue #2029 (#2052)
* Fix the parsing error of Issue #2029

I changed the replace from , -> . to , -> nothing, because the stat got
interpreted as a float and in a later step rounded down to 1

* Addressed the issue in #2029 and the change request in #2052

I changed the code, so that the `,` or `.` only gets deleted if they are
followed by 3 digits. Since artifacts only have a precision of one digit
this method should not delete any `,` or `.` used for floating point
seperation.

* Update libs/gi/art-scanner/src/lib/parse.ts

Add forgotten `\` to escape the `.` of the regex

Co-authored-by: frzyc <frzyc@users.noreply.github.com>

* fixed small bug and ran `yarn run mini-ci`

the regex caught .000 but also everything abouve like .00000 and now it
olny catches, if the `.` is followed by 3 `0`s. I also ran `yarn run
mini-ci` this time.

---------

Co-authored-by: frzyc <frzyc@users.noreply.github.com>