Feature/character card UI (#1873)
* Elementary version of character viewer

This patch makes sure that the data is propagated down to each
individual cards. The patch has pagination on it but the pagination is
still non-functional.

It adds character information in a list format to the main page, based
on data from the database.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* basic implementation for pagination

List pagination function uses slice to leverage c++ (hopefully).

Pagination on the character list view works now, but should be modified
to use infinite scrolling. Or so I've heard.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* implement infinite scrolling for the characters

Infinite scrolling used here is inside a 50% view height box. This to
prevent the page from getting too long when a user has many characters.
There is no styling of the cards yet, as this is up to the owner's
liking.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* remove unused function

Since the function is unused, it can be safely removed. I do think that
it is good practice to use a lib fn for this instead of copying it every
time but it's not up to me.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* fix some linting issues

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* remove null check on character

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* put character inventory inside accordion

The character inventory is now inside an accordion. Nothing else has
changed so far. As per request in PR.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* remove container inside inventory component

The container inside CharacterInventory component made the view a little
narrow. To fit nicely with the other views, removed container from the
character inventory. It is already around the accordion.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* add more stats in the character view

More stats are added based on the PR. I don't play the game so please,
if I butchered any names I aplogize. Correct at your leisure.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* post mini-ci

Some minor beautifications from linter.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* add atk as per request of maintainers.

Adds atk to the character information.

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* post mini-ci fix

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* update submodule from remote

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* revert some changes from bad merge

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>

* remove RelicEditor and RelicInventory

---------

Signed-off-by: Jelte Dirks <j.dirks@outlook.com>
Co-authored-by: frzyc <frzyc@users.noreply.github.com>